+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     the checksum of all the arguments used to create the thumbnail. Using the
     prefix of a thumbnail, it will be possible to identify the original image
     from which it was generated;
+* the `Last-Modified` header is set up and sent to the client. It indicates the
+    date and time at which the thumbnail file was modified for the last time;
 * added `ThumbManager` and `ThumbShell` classes;
 * `resolveFilePath()` method moved from `ThumbCreator` to `ThumbTrait`, because
     this method can be used in different contexts.